+++ NanoXLSX.Formatting.Test/Worksheets/WorksheetTest.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NanoXLSX.Formatting.Test.Worksheets
+{
+    public class WorksheetTest
+    {
+        public void SampleTestMethod()
+        {
+            Workbook workbook = new Workbook();
+           // workbook.CurrentWorksheet.AddCell()
+        }
+    }
+}