+++ NanoXLSX.Formatting/Extensions/FormattedText.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace NanoXLSX
+{
+    /// <summary>
+    /// Class representing a formatted text that can be added to cells
+    /// </summary>
+    public class FormattedText
+    {
+
+    }
+}