+++ NanoXLSX.Formatting/Extensions/WorksheetExtensions.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Text;
+using NanoXLSX.Extensions;
+
+namespace NanoXLSX
+{
+    /// <summary>
+    /// Writer extension methods for the <see cref="NanoXLSX.Worksheet">Workbook</see> class
+    /// </summary>
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    public static class WorksheetExtensions
+    {
+        public static void AddCell(this Worksheet worksheet, FormattedText formattedText, int columnNumber, int rowNumber)
+        {
+
+        }
+    }
+}