+++ NanoXLSX.Formatting/NanoXLSX.Formatting .NET Standard/Class1.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace NanoXLSX.Formatting
-{
-    public class Class1
-    {
-        // DUMMY --> Use source link to .NET library
-    }
-}